grid: Rewrite component segmentation, add union-find algo. Resolves #1
The old component segmentation code had a bunch of duplication and was
hard to reason about. It is replaced with a more general solution.

The old merge code was outright wrong and it caused various weird
issues. It is replaced with union-find, which is an algorithm used to
partition a set of elements into a number of disjoint subsets.

References:
  - https://medium.com/100-days-of-algorithms/day-41-union-find-d0027148376d
  - https://www.youtube.com/watch?v=0jNmHPfA_yE