tui: Change how space bar changes tiles
I think it makes more sense for space to delete anything if it is there,
or create a wire if there is nothing.

That way it's the "delete" tool which happens to also create wire as a
secondary function, instead of being a "wire" tool which happens to also
delete things as a secondary function.